+++ astropy/table/table.py
@@ -2062,3 +2062,8 @@ class QTable(Table):
         If ``col`` is a string then it refers to a column name in this table.
         """
         return super(QTable, self)._is_mixin_column(col, quantity_is_mixin=True)
+
+    def __getstate__(self):
+        columns = dict((key, col if isinstance(col, BaseColumn) else col_copy(col))
+                for key, col in self.columns.items())
+        return (columns, self.meta)